requirements.txt
# Core Framework
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
gradio==4.8.0
accelerate==0.25.0

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# API Integration
groq>=0.33.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.9.1
flake8==6.1.0
